feat: UpdateProductRequest 타입 구현 #2
- 타입 재사용을 위한 ProductResponse 타입 이름 수정
