+++ OpenMarket/OpenMarket/Network/API/Image.swift
@@ -1,0 +1,24 @@
+//
+//  Image.swift
+//  OpenMarket
+//
+//  Created by JeongTaek Han on 2022/01/06.
+//
+
+import Foundation
+
+struct Image: Decodable {
+    let id: Int
+    let url: String
+    let thumbnailURL: String
+    let succeed: Bool
+    let issuedAt: String
+
+    enum CodingKeys: String, CodingKey {
+        case id
+        case url
+        case thumbnailURL = "thumbnail_url"
+        case succeed
+        case issuedAt = "issued_at"
+    }
+}
